javaprep

+++ java/sem 3/mod 2/mod2 progrms/q9.java
@@ -3,12 +3,12 @@ import java.lang.Math;
 
 class shape
 {
-    int dimension;
+    double dimension;
     shape()
     {
         
     }
-    shape(int dm)
+    shape(double dm)
     {
         this.dimension=dm;
         System.out.println(this.dimension);
